refactor: change rate entity prop refactor: recreate rate entity migration
