+++ src/modules/user/dto/rate-request.dto.ts
@@ -5,7 +5,7 @@ import { IsNumber, IsString } from 'class-validator';
 export class RateRequestDto {
   @ApiProperty()
   @IsString()
-  rate_target_id: string;
+  user_target_id: string;
 
   @ApiProperty()
   @Transform(({ value }) => parseInt(value))
